# Improved performance of open tas