--- a/src/Functions.h
+++ b/src/Functions.h
@@ -20,8 +20,8 @@ using std::stringstream;
 
 bool check_input(const string& item, const vector<string>& all_items);
 
-string inline Capitalize_std(const std::string& stringToCapitalize, const bool isItem = false);
-string inline Capitalize(const wxString& stringToCapitalize, const bool isItem = false);
+string Capitalize(const std::string& stringToCapitalize, const bool isItem = false);
+string Capitalize(const wxString& stringToCapitalize, const bool isItem = false);
 
 bool equals_ignore_case(const std::string_view& lhs, const std::string_view& rhs);
 bool starts_with_ignore_case(const std::string& base, const std::string& start);